requirements.txt
dbt-core==1.10.11
dbt-duckdb==1.9.6
pandas==2.3.1
